Experiment for Yesod.
This commit makes lushtags to assume several language extensions are
enabled, regardless of what are actually enabled, making it possible to
create tags for Haskell sources written for Yesod.

The default lushtags could not handle QuasiQuotes, for example, so it
could not parse such sources, which was annoying to me. Ideally, the
extensions enabled for a certain source must be identified from the
source code itself and the cabal configuration file. However, it might
take some time to implement it, so I created an experimental
implementation to see if Haskell-ext can parse Yesod handlers with lots
of QQ code. At this moment, it seems working.